Breakout add party into new page
