+++ src/components/Rsvp.css
@@ -14,7 +14,7 @@
     background-color: white;
 }
 
-#guest-input-wrapper {
+.rsvpWrapper {
     max-width: 400px;
     margin: 2em auto;
 }
